+++ ivy/functional/ivy/layers.py
@@ -18,6 +18,7 @@ from ivy.func_wrapper import (
     handle_device,
     handle_backend_invalid,
 )
+from ivy.functional.ivy.experimental.manipulation import _slice_along_axis
 from ivy.utils.exceptions import handle_exceptions
 
 # Extra #
@@ -2379,6 +2380,108 @@ def lstm_update(
     return ret, (ht, ct)
 
 
+@handle_exceptions
+@handle_nestable
+@handle_array_like_without_promotion
+@inputs_to_ivy_arrays
+@handle_array_function
+def lstm(
+    input: ivy.Array,
+    initial_states: Tuple[ivy.Array],
+    all_weights: Tuple[ivy.Array],
+    has_biases: bool,
+    num_layers: int,
+    dropout: float,
+    train: bool,
+    bidirectional: bool,
+    batch_first: bool = False,
+    batch_sizes: Sequence = None,
+):
+    weights_per_layer = 4 if has_biases else 2
+
+    assert len(all_weights) == num_layers * weights_per_layer * (1 + bidirectional)
+    layer_weights = [
+        all_weights[i : i + weights_per_layer]
+        for i in range(0, len(all_weights), weights_per_layer)
+    ]
+
+    if batch_sizes is not None:
+        input, batch_sizes = _pad_packed_sequence(input, batch_sizes)
+
+    if batch_first:
+        input = ivy.swapaxes(input, 0, 1)
+
+    if dropout and train:
+        raise ivy.utils.exceptions.IvyNotImplementedException()
+
+    unidirectional = not bidirectional
+
+    h0, c0 = initial_states
+    h_outs, c_outs = [], []
+
+    output = input
+    for i in range(num_layers):
+        if unidirectional:
+            if weights_per_layer == 4:
+                weight_ih, weight_hh, (bias_i, bias_h) = _transform_weights(
+                    layer_weights, i
+                )
+            else:
+                weight_ih, weight_hh = _transform_weights_no_bias(layer_weights, i)
+                bias_i = bias_h = None
+
+            state_indices = i, i + 1
+        else:
+            if weights_per_layer == 4:
+                weight_ih_f, weight_hh_f, (bias_i_f, bias_h_f) = _transform_weights(
+                    layer_weights, 2 * i
+                )
+                weight_ih_b, weight_hh_b, (bias_i_b, bias_h_b) = _transform_weights(
+                    layer_weights, 2 * i + 1
+                )
+            else:
+                weight_ih_f, weight_hh_f = _transform_weights_no_bias(
+                    layer_weights, 2 * i
+                )
+                weight_ih_b, weight_hh_b = _transform_weights_no_bias(
+                    layer_weights, 2 * i + 1
+                )
+                bias_i_f = bias_h_f = bias_i_b = bias_h_b = None
+
+            weight_ih = weight_ih_f, weight_ih_b
+            weight_hh = weight_hh_f, weight_hh_b
+            bias_i = bias_i_f, bias_i_b
+            bias_h = bias_h_f, bias_h_b
+
+            state_indices = 2 * i, 2 * i + 2
+
+        output, (h_out, c_out) = _lstm_layer(
+            output,
+            (
+                _retrieve_state(h0, *state_indices, num_layers),
+                _retrieve_state(c0, *state_indices, num_layers),
+            ),
+            (weight_ih, weight_hh),
+            (bias_i, bias_h),
+            bidirectional,
+            batch_first=False,
+            batch_sizes=batch_sizes,
+        )
+        h_outs.append(h_out)
+        c_outs.append(c_out)
+
+    if batch_first:
+        output = ivy.swapaxes(output, 0, 1)
+
+    h_outs = h_out if num_layers == 1 else ivy.concat(h_outs, axis=0)
+    c_outs = c_out if num_layers == 1 else ivy.concat(c_outs, axis=0)
+
+    if batch_sizes is not None:
+        output = _pack_padded_sequence(output, batch_sizes)[0]
+
+    return output, h_outs, c_outs
+
+
 # Helpers #
 
 
@@ -2621,6 +2724,131 @@ def _convert_boxes_to_roi_format(boxes):
     ids = ivy.concat(temp, axis=0)
     rois = ivy.concat([ids, concat_boxes], axis=1)
     return rois
+
+
+def _lstm_cell(
+    x,
+    init_h,
+    init_c,
+    kernel,
+    recurrent_kernel,
+    bias,
+    recurrent_bias,
+    batch_first,
+    batch_sizes=None,
+):
+    init_h = ivy.squeeze(init_h, axis=0)
+    init_c = ivy.squeeze(init_c, axis=0)
+    out, states = lstm_update(
+        x,
+        init_h,
+        init_c,
+        kernel,
+        recurrent_kernel,
+        bias=bias,
+        recurrent_bias=recurrent_bias,
+        time_major=not batch_first,
+    )
+    h, c = states
+    h = ivy.expand_dims(h) if len(h.shape) == 2 else h
+    c = ivy.expand_dims(c) if len(c.shape) == 2 else c
+    return out, (h, c)
+
+
+def _lstm_layer(
+    x, hidden, weights, biases, bidirectional, batch_first, batch_sizes=None
+):
+    if not bidirectional:
+        result, (h, c) = _lstm_cell(
+            x,
+            *hidden,
+            *weights,
+            *biases,
+            batch_first=batch_first,
+            batch_sizes=batch_sizes,
+        )
+    else:
+        result_fw, (h_fw, c_fw) = _lstm_cell(
+            x,
+            hidden[0][:1],
+            hidden[1][:1],
+            weights[0][0],
+            weights[1][0],
+            biases[0][0],
+            biases[1][0],
+            batch_first=batch_first,
+            batch_sizes=batch_sizes,
+        )
+        x_reversed = ivy.flip(x, axis=0)
+        result_bw, (h_bw, c_bw) = _lstm_cell(
+            x_reversed,
+            hidden[0][1:],
+            hidden[1][1:],
+            weights[0][1],
+            weights[1][1],
+            biases[0][1],
+            biases[1][1],
+            batch_first=batch_first,
+            batch_sizes=batch_sizes,
+        )
+        result_bw = ivy.flip(result_bw, axis=0)
+        result = ivy.concat([result_fw, result_bw], axis=len(result_fw.shape) - 1)
+        c = ivy.concat([c_fw, c_bw], axis=0)
+        h = ivy.concat([h_fw, h_bw], axis=0)
+    return result, (h, c)
+
+
+def _pack_padded_sequence(input, lengths):
+    input = ivy.swapaxes(input, 0, 1)
+    data = []
+    batch_sizes = []
+    for i in range(int(max(lengths))):
+        valid_data_mask = ivy.array(lengths) > i
+        data.append(input[valid_data_mask, i])
+        batch_sizes.append(int(sum(valid_data_mask)))
+    data = ivy.concat(data)
+    batch_sizes = ivy.array(batch_sizes, dtype=ivy.int64)
+    return data, batch_sizes
+
+
+def _pad_packed_sequence(data, batch_sizes):
+    padded_data = ivy.full(
+        (len(batch_sizes), int(max(batch_sizes)), *data.shape[1:]),
+        0,
+        dtype=data.dtype,
+        device=data.device,
+    )
+    data_offset = 0
+    for i, batch_size in enumerate(batch_sizes):
+        batch_size = int(batch_size)
+        padded_data[i, :batch_size] = data[data_offset : data_offset + batch_size]
+        data_offset += batch_size
+    lengths = ivy.sum(
+        ivy.arange(1, int(max(batch_sizes)) + 1)[:, ivy.newaxis] <= batch_sizes,
+        axis=1,
+        dtype=ivy.int64,
+    )
+    return padded_data, lengths
+
+
+def _retrieve_state(x, start, end, num_layers):
+    return x if num_layers == 1 else _slice_along_axis(x, start=start, stop=end, axis=0)
+
+
+def _transform_weights(layer_weights, layer_index):
+    weights = layer_weights[layer_index]
+    weight_ih, weight_hh, bias_ih, bias_hh = weights
+    return (
+        ivy.swapaxes(weight_ih, 0, 1),
+        ivy.swapaxes(weight_hh, 0, 1),
+        (bias_ih, bias_hh),
+    )
+
+
+def _transform_weights_no_bias(layer_weights, layer_index):
+    weights = layer_weights[layer_index]
+    weight_ih, weight_hh = weights
+    return ivy.swapaxes(weight_ih, 0, 1), ivy.swapaxes(weight_hh, 0, 1)
 
 
 @handle_exceptions
